renames execution_models to models
