+++ openvair/libs/cli/executor.py
@@ -12,12 +12,12 @@ import os
 from typing import Dict, List
 from subprocess import PIPE, Popen, TimeoutExpired
 
+from openvair.libs.cli.models import ExecuteParams, ExecutionResult
 from openvair.libs.cli.exceptions import (
     ExecuteError,
     ExecuteTimeoutExpiredError,
 )
 from openvair.modules.tools.utils import LOG
-from openvair.libs.cli.execution_models import ExecuteParams, ExecutionResult
 
 
 def __terminate_process(proc: Popen, cmd_str: str) -> str:
